update: only user can edit

+++ comments/views.py
@@ -2,6 +2,7 @@
 
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from .models import Comment
 from .forms import CommentForm
 from events.models import Event
@@ -25,6 +26,10 @@ def add_comment(request, event_id):
 @login_required
 def edit_comment(request, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
+    event = comment.event  # イベントを取得
+    comments = Comment.objects.filter(event=event)  # コメントを取得
+    if comment.created_by != request.user:
+        return render(request, 'events/event_detail.html', {'event': event, 'comments': comments, 'error_message': '他の人の投稿は編集できません'})
     if request.method == 'POST':
         form = CommentForm(request.POST, instance=comment)
         if form.is_valid():
@@ -38,6 +43,10 @@ def edit_comment(request, comment_id):
 @login_required
 def delete_comment(request, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
+    event = comment.event  # イベントを取得
+    comments = Comment.objects.filter(event=event)  # コメントを取得
+    if comment.created_by != request.user:
+        return render(request, 'events/event_detail.html', {'event': event, 'comments': comments, 'error_message': '他の人の投稿は削除できません'})
     if request.method == 'POST':
         event_id = comment.event.id
         comment.delete()
